LimitedWordsBot/requirements.txt
discord.py
termcolor
firebase_admin